SystemCleaner: Improve exclusion matching
Better handle exclusions that first match during scan, but then other "parent" folders of an exclusion still pass and need to be removed from the final filter contents too.
